B Tennis' ball goes way off the grid
When the face is in a 'show no seconds' mode for at least a few seconds,
then the ball - and with it the rackets - will be way of the screen, as
the movement is determined by the time elapsed.

This change puts the ball on the border of the visible area when it is
outside.